pyproject.toml
[project]
name = "cisco-contract-intelligence"
version = "1.0.0"
description = "AI-powered contract analysis for Cisco CDP"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "anthropic>=0.50.0",
    "pdfplumber>=0.11.0",
    "pytesseract>=0.3.13",
    "Pillow>=11.0.0",
    "pdf2image>=1.17.0",
    "python-multipart>=0.0.17",
    "aiosqlite>=0.20.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
contract-cli = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
